requirements.txt
asgiref==3.8.1
Django==5.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.9
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
